split out androgee into multiple files to make him more extensible
