--- run-black-hook.sh
@@ -1,1 +1,0 @@
-poetry run black androgee/androgee.py
